moved clients to auth
